fix: extract duplicated InternalAuthFilter into shared atlas-common library
