+++ services/atlas-common/src/main/java/com/atlas/common/security/CommonSecurityAutoConfiguration.java
@@ -1,0 +1,9 @@
+package com.atlas.common.security;
+
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+
+@AutoConfiguration
+@ComponentScan(basePackageClasses = InternalAuthFilter.class)
+public class CommonSecurityAutoConfiguration {
+}
