+++ services/atlas-common/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.atlas.common.security.CommonSecurityAutoConfiguration
